[spacebar] remove leading space/zero from day
`%e` is the day of the month with a space as padding (so that it is two
characters), while `%d` is the same but with a leading zero for the same
goal. In either case, we don’t want two characters and it turns out you
can add a hyphen to remove the padded zero.