[yabai] add proper stacking shortcuts
`alt - s` creates a stack from the adjacent window. Not ideal IMO; I’d
prefer if it created it using whatever windows were in the space that
`zoom-parent` covered.

The same focus navigation commands for up and down are used for
navigating the stack. I could do this for right and left as well, so
that it would be like tabbed layouts in sway, but I found this odd.