[git] add commit triage alias
What is a “commit triage?” It’s a phrase I use for when your work tree
is full of stuff you’ve got to break into separate commits and you are
too lazy to write the git commit command each time. It runs until the
work tree is has no unmodified or staged files.